+++ src/file_handler.h
@@ -5,7 +5,7 @@
 #define DELIMITER ","
 #define BLOCK_SIZE (4 * 1024 * 1024)  // 4MB
 
-int getstarsfromfile(char *dirname, Star *estrellas);
+unsigned long getstarsfromfile(char *dirname, Star *estrellas);
 void free_stars(Star *stars);
 
 #endif // FILE_HANDLER_H
